Optionally only trigger visible cues in Cart layout
Theatre shows often have more than 26 cues, which means that is isn't possible
to trigger them individually from hotkeys. Stepping through them in list mode
is an option, but this is adds complexity if things deviate from the script,
i.e. exactly when you don't want complexity.

Add an option to scope triggers to the currently visible page in the Cart view.
This is off by default and saved with the show itself.

This resolves https://github.com/FrancescoCeruti/linux-show-player/issues/328

Signed-off-by: Phil Wise <phil@phil-wise.com>